requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django==4.1.1
django-decouple==2.1
django-extensions==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1; python_version >= '3.6'
pytz==2022.2.1
sqlparse==0.4.2; python_version >= '3.5'
